requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-modeltranslation==0.19.11
django-ninja==1.3.0
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
